Refactored the header. Split components, and functions into external files for imports. Display an empty fragment for fetching the firestore user data instead of login/signup when the user is already signed in.
