+++ src/containers/Header/HeaderStyledComponents.js
@@ -81,3 +81,6 @@ export const MobileNavToggler = styled.span`
     color: #0d238e;
   }
 `;
+
+export const Link = styled.span`
+`
