+++ src/containers/Header/MobileNavMenu.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Link } from "react-router-dom";
+// import { Link } from "react-router-dom";
 import * as ROUTES from "../../routes/constants/routes";
 import * as Nav from './HeaderStyledComponents';
 import CartDropdown from "./CartDropdown/CartDropdown";
@@ -20,22 +20,22 @@ const MobileNavMenu = (props) => {
           <Nav.MobileNavContainer>
             <Nav.MobileNavList style={{ listStyleType: "none" }}>
               <Nav.MobileNavListItem>
-                <Link onClick={e => gotolink(e, ROUTES.HOME)}>Home</Link>
+                <Nav.Link onClick={e => gotolink(e, ROUTES.HOME)}>Home</Nav.Link>
               </Nav.MobileNavListItem>
               <Nav.MobileNavListItem>
-                <Link onClick={e => gotolink(e, ROUTES.HOME)}>About</Link>
+                <Nav.Link onClick={e => gotolink(e, ROUTES.HOME)}>About</Nav.Link>
               </Nav.MobileNavListItem>
               <Nav.MobileNavListItem>
-                <Link onClick={e => gotolink(e, ROUTES.DISCOVER_HOMEOWNER)}>
+                <Nav.Link onClick={e => gotolink(e, ROUTES.DISCOVER_HOMEOWNER)}>
                   Get Your FREE Flood Score
-                </Link>
+                </Nav.Link>
               </Nav.MobileNavListItem>
               {firestoreUser ? (
                 <>
                   <Nav.MobileNavListItem>
-                    <Link onClick={e => gotolink(e, ROUTES.ACCOUNT_DASHBOARD)}>
+                    <Nav.Link onClick={e => gotolink(e, ROUTES.ACCOUNT_DASHBOARD)}>
                       Hi, {firestoreUser.firstName}
-                    </Link>
+                    </Nav.Link>
                   </Nav.MobileNavListItem>
                   <Nav.MobileNavListItem onClick={() => setShowMobileNav(false)}>
                     <CartDropdown />
@@ -44,14 +44,14 @@ const MobileNavMenu = (props) => {
               ) : (
                 <>
                   <Nav.MobileNavListItem>
-                    <Link onClick={e => gotolink(e, ROUTES.SIGN_IN)}>
+                    <Nav.Link onClick={e => gotolink(e, ROUTES.SIGN_IN)}>
                       Login
-                    </Link>
+                    </Nav.Link>
                   </Nav.MobileNavListItem>
                   <Nav.MobileNavListItem>
-                    <Link onClick={e => gotolink(e, ROUTES.CHECKOUT_FREE)}>
+                    <Nav.Link onClick={e => gotolink(e, ROUTES.CHECKOUT_FREE)}>
                       Sign Up
-                    </Link>
+                    </Nav.Link>
                   </Nav.MobileNavListItem>
                   <Nav.MobileNavListItem onClick={() => setShowMobileNav(false)}>
                     <CartDropdown />
